feat: validade different authorId

+++ src/domain/forum/application/use-cases/delete-question-use-case.ts
@@ -1,6 +1,7 @@
 import { QuestionRepository } from '../repositories/question-repository'
 
 interface DeleteQuestionUseCaseRequest {
+  authorId: string
   questionId: string
 }
 
@@ -10,12 +11,17 @@ export class DeleteQuestionUseCase {
   constructor(private questionRepository: QuestionRepository) {}
 
   async execute({
+    authorId,
     questionId,
   }: DeleteQuestionUseCaseRequest): Promise<DeleteQuestionUseCaseResponse> {
     const question = await this.questionRepository.findById(questionId)
 
     if (!question) {
       throw new Error('Question not found')
+    }
+
+    if (authorId !== question.authorId.toString()) {
+      throw new Error('Not authorized')
     }
 
     await this.questionRepository.delete(question)
